+++ src/Views/Filters/DateFilter.php
@@ -7,6 +7,8 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 
 class DateFilter extends Filter
 {
+    public string $viewPath = 'livewire-tables::components.tools.filters.date';
+    
     public function config(array $config = []): DateFilter
     {
         $this->config = [...config('livewire-tables.dateFilter.defaultConfig'), ...$config];
@@ -46,7 +48,7 @@ class DateFilter extends Filter
     }
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
-    {
-        return view('livewire-tables::components.tools.filters.date', $this->getFilterDisplayData());
+    {        
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }