+++ src/Views/Filters/DateRangeFilter.php
@@ -7,6 +7,7 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 class DateRangeFilter extends Filter
 {
     public array $options = [];
+    public string $viewPath = 'livewire-tables::components.tools.filters.date-range';
 
     public function config(array $config = []): DateRangeFilter
     {
@@ -190,6 +191,6 @@ class DateRangeFilter extends Filter
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        return view('livewire-tables::components.tools.filters.date-range', $this->getFilterDisplayData());
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }