+++ src/Views/Filters/MultiSelectDropdownFilter.php
@@ -7,6 +7,7 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 class MultiSelectDropdownFilter extends Filter
 {
     public array $options = [];
+    public string $viewPath = 'livewire-tables::components.tools.filters.multi-select-dropdown';
 
     protected string $firstOption = '';
 
@@ -107,6 +108,6 @@ class MultiSelectDropdownFilter extends Filter
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        return view('livewire-tables::components.tools.filters.multi-select-dropdown', $this->getFilterDisplayData());
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }