+++ src/Views/Filters/MultiSelectFilter.php
@@ -7,6 +7,7 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 class MultiSelectFilter extends Filter
 {
     public array $options = [];
+    public string $viewPath = 'livewire-tables::components.tools.filters.multi-select';
 
     protected string $firstOption = '';
 
@@ -96,6 +97,6 @@ class MultiSelectFilter extends Filter
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        return view('livewire-tables::components.tools.filters.multi-select', $this->getFilterDisplayData());
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }