+++ src/Views/Filters/NumberRangeFilter.php
@@ -7,6 +7,7 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 class NumberRangeFilter extends Filter
 {
     public array $options = [];
+    public string $viewPath = 'livewire-tables::components.tools.filters.number-range';
 
     public function options(array $options = []): NumberRangeFilter
     {
@@ -88,6 +89,6 @@ class NumberRangeFilter extends Filter
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        return view('livewire-tables::components.tools.filters.number-range', $this->getFilterDisplayData());
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }