+++ src/Views/Traits/Helpers/FilterHelpers.php
@@ -265,4 +265,10 @@ trait FilterHelpers
     {
         return array_merge($this->getGenericDisplayData(), ['filter' => $this]);
     }
+
+
+    public function getViewPath():
+    {
+        return $this->viewPath;
+    }
 }